+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `pipe` function
 * Add `match` function
 * Add `startsWith` function
+* Add `endsWith` function
 
 ## 3.2.0 (2019-04-23)